migration error fixed

+++ database/migrations/2023_10_10_080000_create_verifications_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('verifications', function (Blueprint $table) {
             $table->id();
-            $table->BigInteger('user_id');
+            $table->unsignedBigInteger('user_id');
             $table->string('first_name');
             $table->string('last_name');
             $table->string('date_of_birth');
